chore(infra): 旧Bedrockディレクトリの不要部分を削除しバレルを廃止\n\n- 削除: external/bedrock/ContentProcessing, QueryOptimization（配下実装とテスト）\n- 削除: external/bedrock/index.ts（空のバレル）\n- Adapter/Factoryへ完全移行、全テスト緑\n\nRefs: #13
