--- src/shared/infrastructure/external/bedrock/ContentProcessing/index.test.ts
@@ -1,12 +1,0 @@
-import { describe, it, expect } from "vitest";
-
-describe("llm/adapters/bedrock/contentProcessingAdapter (moved)", () => {
-  it("BedrockContentProcessingAdapter をエクスポートする", async () => {
-    const mod = (await import(
-      "@/shared/infrastructure/external/llm/adapters/bedrock/contentProcessingAdapter"
-    )) as unknown as {
-      BedrockContentProcessingAdapter: unknown;
-    };
-    expect(typeof mod.BedrockContentProcessingAdapter).toBe("function");
-  });
-});
